refactor(fsd): Приведение проекта к большему соответствию fsd структуре.
- перенёс user-срез (например  isLoggedIn:) из app/model в entities/user/model, appSlice теперь содержит только глобальное состояние.

- Удалена лишняя обертка было shared/ui/button/ui/button.tsx стало shared/ui/button/button.tsx, настроен реэкспорт

- добавил слой widgets

- перенос assets в дир. shared

- Исправлены импорты в форме входа
