+++ app/model/appSlice.ts
@@ -3,22 +3,23 @@ import { createSlice } from "@reduxjs/toolkit"
 export const appSlice = createSlice({
   name: "app",
   initialState: {
-    isLoggedIn: false,
+    themeMode: "dark", //заглушка
+    error: null as Error,
   },
   selectors: {
-    selectIsLoggedIn: (state) => state.isLoggedIn,
+    selectThemeMode: (state) => state.themeMode,
   },
   reducers: (create) => ({
-    setIsLoggedIn: create.reducer<{ isLoggedIn: boolean }>((state, action) => {
-      state.isLoggedIn = action.payload.isLoggedIn
+    changeTheme: create.reducer<{ themeMode: string }>((state, action) => {
+      state.themeMode = action.payload.themeMode
     }),
   }),
 })
 
-export const { setIsLoggedIn } = appSlice.actions
-export const { selectIsLoggedIn } = appSlice.selectors
+export const { changeTheme } = appSlice.actions
+export const { selectThemeMode } = appSlice.selectors
 export const appReducer = appSlice.reducer
 
 //types
 export type Error = null | string
-export type AppInitialState = ReturnType<typeof appSlice.getInitialState>
+// export type AppInitialState = ReturnType<typeof appSlice.getInitialState>
