+++ features/auth/ui/Register/InfoForm/InfoForm.module.scss
@@ -19,11 +19,6 @@
   margin-bottom: 72px;
 }
 
-.form {
-  display: flex;
-  flex-direction: column;
-}
-
 .input {
   margin-bottom: 24px;
 }
