+++ shared/ui/index.tsx
@@ -1,0 +1,2 @@
+export { Button } from "./button/button"
+// export { Checkbox } from "./checkbox/checkbox"
